requirements.txt
Django==4.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
django-cors-headers==3.11.0
whitenoise==6.2.0
dj-database-url==0.5.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
djangorestframework-jwt
djangorestframework-simplejwt